fix(workspace): move runner-replay to dir
