+++ packages/runner/user-flow-replay/package.json
@@ -1,0 +1,4 @@
+{
+  "name": "@app-speed/runner-user-flow-replay",
+  "version": "0.0.1"
+}
