+++ packages/runner/user-flow-replay/project.json
@@ -1,7 +1,7 @@
 {
-  "name": "user-flow-replay",
-  "$schema": "../../node_modules/nx/schemas/project-schema.json",
-  "sourceRoot": "packages/user-flow-replay/src",
+  "name": "runner-user-flow-replay",
+  "$schema": "../../../node_modules/nx/schemas/project-schema.json",
+  "sourceRoot": "packages/runner/user-flow-replay/src",
   "projectType": "library",
   "tags": [],
   "targets": {
@@ -9,10 +9,10 @@
       "executor": "@nx/esbuild:esbuild",
       "outputs": ["{options.outputPath}"],
       "options": {
-        "outputPath": "dist/packages/user-flow-replay",
-        "main": "packages/user-flow-replay/public-api.ts",
-        "tsConfig": "packages/user-flow-replay/tsconfig.lib.json",
-        "assets": ["packages/user-flow-replay/*.md"],
+        "outputPath": "dist/packages/runner/user-flow-replay",
+        "main": "packages/runner/user-flow-replay/public-api.ts",
+        "tsConfig": "packages/runner/user-flow-replay/tsconfig.lib.json",
+        "assets": ["packages/runner/user-flow-replay/*.md"],
         "generatePackageJson": true
       }
     },
@@ -30,7 +30,7 @@
       ],
       "options": {
         "passWithNoTests": true,
-        "coverage.reportsDirectory": "../../coverage/packages/user-flow-replay"
+        "coverage.reportsDirectory": "../../coverage/packages/runner/user-flow-replay"
       }
     }
   }
