+++ packages/runner/user-flow-replay/public-api.ts
@@ -1,0 +1,1 @@
+export { UserFlowAudit } from './src/user-flow-replay';
