+++ packages/runner/user-flow-replay/src/parse.ts
@@ -27,7 +27,7 @@ export function parse(recordingJson: any): AppSpeedUserFlow {
   const parsed: AppSpeedUserFlow = puppeteerReplayParse({ ...recordingJson, steps });
 
   ufArr.forEach((value, index) => {
-    value && (parsed.steps.splice(index, 0, value));
+    value && parsed.steps.splice(index, 0, value);
   });
 
   // parse customEvents from our stringify function
