+++ packages/runner/user-flow-replay/src/test-data/puppeteer-replay.json
@@ -13,42 +13,21 @@
     {
       "type": "click",
       "target": "main",
-      "selectors": [
-        [
-          "aria/Mocha"
-        ],
-        [
-          "[data-test=Mocha]"
-        ]
-      ],
+      "selectors": [["aria/Mocha"], ["[data-test=Mocha]"]],
       "offsetY": 144.24940490722656,
       "offsetX": 186.73519897460938
     },
     {
       "type": "click",
       "target": "main",
-      "selectors": [
-        [
-          "aria/Proceed to checkout"
-        ],
-        [
-          "[data-test=checkout]"
-        ]
-      ],
+      "selectors": [["aria/Proceed to checkout"], ["[data-test=checkout]"]],
       "offsetY": 17.7916259765625,
       "offsetX": 146.8541259765625
     },
     {
       "type": "change",
       "value": "test@test.at",
-      "selectors": [
-        [
-          "aria/Name"
-        ],
-        [
-          "#name"
-        ]
-      ],
+      "selectors": [["aria/Name"], ["#name"]],
       "target": "main"
     }
   ]
