+++ packages/runner/user-flow-replay/src/user-flow-replay.ts
@@ -34,10 +34,10 @@ export class UserFlowAudit {
   }
 
   private async initializeRunnerContext(): Promise<UserFlowRunnerContext> {
-    const browser = await launch({ headless: "new", args: ['--no-sandbox', '--disable-setuid-sandbox'] });
+    const browser = await launch({ headless: 'new', args: ['--no-sandbox', '--disable-setuid-sandbox'] });
     const page = await browser.newPage();
     const flow = await startFlow(page, { name: this.replayScript.title });
-    return {browser, page, flow};
+    return { browser, page, flow };
   }
 
   private async createRunner(runnerContext: UserFlowRunnerContext): Promise<Runner> {
@@ -53,7 +53,7 @@ export class UserFlowAudit {
   private async extractResults(runnerContext: UserFlowRunnerContext): Promise<ResultReports> {
     const jsonReport = JSON.stringify(await runnerContext.flow.createFlowResult());
     const htmlReport = await runnerContext.flow.generateReport();
-    return {jsonReport, htmlReport};
+    return { jsonReport, htmlReport };
   }
 
   private async terminateRunnerContext(runnerContext: UserFlowRunnerContext): Promise<void> {
