+++ packages/runner/user-flow-replay/tsconfig.json
@@ -1,5 +1,5 @@
 {
-  "extends": "../../tsconfig.base.json",
+  "extends": "../../../tsconfig.base.json",
   "compilerOptions": {
     "module": "es2022",
     "forceConsistentCasingInFileNames": true,
