--- packages/user-flow-replay/package.json
@@ -1,4 +1,0 @@
-{
-  "name": "@app-speed/user-flow-replay",
-  "version": "0.0.1"
-}
