--- packages/user-flow-replay/public-api.ts
@@ -1,1 +1,0 @@
-export {UserFlowAudit} from './src/user-flow-replay';
